requirements.txt
flask
passlib
pyodbc
python-dotenv
requests
gunicorn
SQLAlchemy

azure-ai-formrecognizer>=3.3.0
azure-storage-blob